feat: Add delete user endpoint
This commit adds a new endpoint to the SecurityController that allows deleting a user. The endpoint is implemented in both the API Gateway and Security Service. The API Gateway sends a DELETE request to the Security Service to delete the user with the specified ID. If the request is successful, the user is deleted from the database and a 204 No Content response is returned. If the request fails, an error response with appropriate error messages is returned.