feat: Add update file endpoint to FileManagementController
This commit adds a new endpoint to the FileManagementController that allows updating a file. The endpoint is implemented in both the API Gateway and File Management Service. The API Gateway sends a POST request to the File Management Service to update the file with the specified ID. If the request is successful, the file is updated in the database and a 201 Created response is returned. If the request fails, an error response with appropriate error messages is returned.